+++ src/scenarios.rs
@@ -415,7 +415,7 @@ impl JumpToBall {
 
         #[cfg(feature = "enable_stats")]
         {
-            ctx.stats.play_micro_ticks += ctx.near_micro_ticks_per_tick as i32;
+            ctx.stats.play_micro_ticks += ctx.near_micro_ticks_per_tick;
             ctx.stats.game_micro_ticks += ctx.near_micro_ticks_per_tick;
         }
